feat: restructuring code, adjusting alarm, lb rule path, lb rule priority + addressing other minor comments
